Use Athena to build PDF, if Docker installed
Merges https://github.com/manubot/rootstock/pull/210
Closes https://github.com/manubot/rootstock/issues/170

More information about Athena is available at
https://github.com/arachnys/athenapdf

Based on meta-review commit:
https://github.com/greenelab/meta-review/commit/cffcfeaea0b6859051d16ab1e263d8bcf62a24b1

+++ build/build.sh
@@ -50,9 +50,12 @@ pandoc --verbose \
   --output=output/manuscript.html \
   $INPUT_PATH
 
+# Return null if docker command is missing, otherwise return path to docker
+DOCKER_EXISTS=`command -v docker`
+
 # Create PDF output (unless BUILD_PDF environment variable equals "false")
-if [ "$BUILD_PDF" != "false" ]; then
-  echo "Exporting PDF manuscript"
+if [ "$BUILD_PDF" != "false" ] && [ -z "$DOCKER_EXISTS" ]; then
+  echo "Exporting PDF manuscript using WeasyPrint"
   if [ -L images ]; then rm images; fi  # if images is a symlink, remove it
   ln -s content/images
   pandoc \
@@ -71,6 +74,22 @@ if [ "$BUILD_PDF" != "false" ]; then
     --output=output/manuscript.pdf \
     $INPUT_PATH
   rm images
+fi
+
+# Create PDF output (unless BUILD_PDF environment variable equals "false")
+if [ "$BUILD_PDF" != "false" ] && [ -n "$DOCKER_EXISTS" ]; then
+  echo "Exporting PDF manuscript using Docker + Athena"
+  if [ -d output/images ]; then rm -rf output/images; fi  # if images is a directory, remove it
+  cp -R -L content/images output/
+  docker run \
+    --rm \
+    --volume `pwd`/output:/converted/ \
+    --security-opt seccomp:unconfined \
+    arachnysdocker/athenapdf:2.16.0 \
+    athenapdf \
+    --delay=2000 \
+    manuscript.html manuscript.pdf
+  rm -rf output/images
 fi
 
 # Create DOCX output (if BUILD_DOCX environment variable equals "true")
